requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Session==0.5.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.26.1
opencv-python==4.5.5.62
Pillow==10.1.0
protobuf==4.21.12
Werkzeug==3.0.0
zipp==3.17.0